// apps/cofepris/src/app/application/tramites/260910/models/mercancia.model.ts
export interface Mercancia {
  /** Clasificación del producto al que pertenece la mercancía. */
  clasificaionProductos: string;

  /** Identificador numérico para especificar el producto. */
  especificarProducto: number;

  /** Nombre específico del producto. */
  nombreProductoEspecifico: string;

  /** Distintiva asociada al producto. */
  distintiva: string;

  /** Cientifico asociada al producto. */
  cientifico: string;

  /** Tipo de producto representado por un identificador numérico. */
  tipoProducto: number;

  /** Forma farmacéutica */
  farmaceutica: string;

  /** Estado físico */
  fisico: string;

  /** Fracción arancelaria de la mercancía. */
  fraccionArancelaria: string;

  /** Descripción detallada de la fracción arancelaria. */
  descripcionFraccionArancelaria: string;

  /** Cantidad de mercancía medida en la Unidad de Medida de Tarifa (UMT). */
  cantidadUMT: string;

  /** Unidad de Medida de Tarifa (UMT) utilizada para medir la cantidad. */
  umt: string;

  /** Cantidad de mercancía medida en la Unidad de Medida de Comercialización (UMC). */
  cantidadUMC: string;

  /** Unidad de Medida de Comercialización (UMC) utilizada para medir la cantidad. */
  umc: number;

  /** Presentación Farmacéutica o tipo de envase */
  presentacionFarmaceutica: string;

  /** Registro sanitario asociado a la mercancía. */
  registroSanitario: number;

  /**
   * Fecha de caducidad de la mercancía.
   */
  fechaCaducidad: string;

  /** País de origen de la mercancía. */
  paisDeOrigen: string;

  /** País de procedencia de la mercancía. */
  paisDeProcedencia: string;

  /** Uso específico de la mercancía. */
  usoEspecifico: string;
}